bettingbot/requirements.txt
attrs==21.2.0
Automat==20.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
google-api-core==1.31.0
google-auth==1.32.1
google-cloud-core==1.7.1
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
iniconfig==1.1.1
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
packaging==21.0
parsel==1.6.0
pluggy==0.13.1
priority==1.3.0
Protego==0.1.16
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
queuelib==1.6.1
requests==2.25.1
rsa==4.7.2
Scrapy==2.5.0
service-identity==21.1.0
six==1.16.0
toml==0.10.2
Twisted==21.2.0
urllib3==1.26.6
w3lib==1.22.0
zope.interface==5.4.0
Flask==1.0.2
gunicorn==19.9.0